1st pass on renaming weapons
