+++ ark_replacer_a3/cfgWeapons.hpp
@@ -1,0 +1,45 @@
+class Mode_SemiAuto;
+class Mode_FullAuto;
+
+class CfgWeapons {
+    class arifle_SPAR_01_base_F;
+    class arifle_SPAR_01_blk_F: arifle_SPAR_01_base_F {
+        displayName = STR_ARK_arifle_SPAR_01_blk;
+    };
+    class arifle_SPAR_01_khk_F: arifle_SPAR_01_base_F {
+        displayName = STR_ARK_arifle_SPAR_01_khk;
+    };
+    class arifle_SPAR_01_snd_F: arifle_SPAR_01_base_F {
+        displayName = STR_ARK_arifle_SPAR_01_snd;
+    };
+    class arifle_SPAR_01_GL_base_F;
+    class arifle_SPAR_01_GL_blk_F: arifle_SPAR_01_GL_base_F {
+        displayName = STR_ARK_arifle_SPAR_01_GL_blk;
+    };
+    class arifle_SPAR_01_GL_khk_F: arifle_SPAR_01_GL_base_F {
+        displayName = STR_ARK_arifle_SPAR_01_GL_khk);
+    };
+    class arifle_SPAR_01_GL_snd_F: arifle_SPAR_01_GL_base_F {
+        displayName = STR_ARK_arifle_SPAR_01_GL_snd);
+    };
+    class arifle_SPAR_02_base_F;
+    class arifle_SPAR_02_blk_F: arifle_SPAR_02_base_F {
+        displayName = STR_ARK_arifle_SPAR_02_blk);
+    };
+    class arifle_SPAR_02_khk_F: arifle_SPAR_02_base_F {
+        displayName = STR_ARK_arifle_SPAR_02_khk);
+    };
+    class arifle_SPAR_02_snd_F: arifle_SPAR_02_base_F {
+        displayName = STR_ARK_arifle_SPAR_02_snd);
+    };
+    class arifle_SPAR_03_base_F;
+    class arifle_SPAR_03_blk_F: arifle_SPAR_03_base_F {
+        displayName = STR_ARK_arifle_SPAR_03_blk);
+    };
+    class arifle_SPAR_03_khk_F: arifle_SPAR_03_base_F {
+        displayName = STR_ARK_arifle_SPAR_03_khk);
+    };
+    class arifle_SPAR_03_snd_F: arifle_SPAR_03_base_F {
+        displayName = STR_ARK_arifle_SPAR_03_snd);
+    };
+};
