+++ ark_replacer_a3/stringtable.xml
@@ -1,0 +1,161 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project name="ARK_Replacer">
+    <Package name="weapon_names">
+        <Key ID="STR_ARK_arifle_SPAR_01_blk">
+            <English>HK416A5 10" (Black)</English>
+            <Czech>HK416A5 10" (Cerná)</Czech>
+            <French>HK416A5 10" (Noir)</French>
+            <Spanish>HK416A5 10" (Negro)</Spanish>
+            <Russian>HK416A5 10" (??????)</Russian>
+            <Polish>HK416A5 10" (czarny)</Polish>
+            <German>HK416A5 10" (Schwarz)</German>
+            <Italian>HK416A5 10" (Nero)</Italian>
+            <Hungarian>HK416A5 10" (Fekete)</Hungarian>
+            <Portuguese>HK416A5 10" (Preto)</Portuguese>
+            <Japanese>HK416A5 10" (?)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_01_khk">
+            <English>HK416A5 10" (Khaki)</English>
+            <Czech>HK416A5 10" (Khaki)</Czech>
+            <French>HK416A5 10" (Kaki)</French>
+            <Spanish>HK416A5 10" (Caqui)</Spanish>
+            <Russian>HK416A5 10" (????)</Russian>
+            <Polish>HK416A5 10" (khaki)</Polish>
+            <German>HK416A5 10" (Khaki)</German>
+            <Italian>HK416A5 10" (Khaki)</Italian>
+            <Hungarian>HK416A5 10" (Khaki)</Hungarian>
+            <Portuguese>HK416A5 10" (Caqui)</Portuguese>
+            <Japanese>HK416A5 10" (??)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_01_snd">
+            <English>HK416A5 10" (Sand)</English>
+            <Czech>HK416A5 10" (Písková)</Czech>
+            <French>HK416A5 10" (Beige)</French>
+            <Spanish>HK416A5 10" (Arena)</Spanish>
+            <Russian>HK416A5 10" (????????)</Russian>
+            <German>HK416A5 10" (Sand)</German>
+            <Polish>HK416A5 10" (piaskowy)</Polish>
+            <Italian>HK416A5 10" (Sabbia)</Italian>
+            <Hungarian>HK416A5 10" (Homok)</Hungarian>
+            <Portuguese>HK416A5 10" (Deserto)</Portuguese>
+            <Japanese>HK416A5 10" (??)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_01_GL_blk">
+            <English>HK416A5 10" GL (Black)</English>
+            <Czech>HK416A5 10" GL (Cerná)</Czech>
+            <French>HK416A5 10" GL (Noir)</French>
+            <Spanish>HK416A5 10" GL (Negro)</Spanish>
+            <Russian>HK416A5 10" GL (??????)</Russian>
+            <Polish>HK416A5 10" GL (czarny)</Polish>
+            <German>HK416A5 10" GL (Schwarz)</German>
+            <Italian>HK416A5 10" GL (Nero)</Italian>
+            <Hungarian>HK416A5 10" GL (Fekete)</Hungarian>
+            <Portuguese>HK416A5 10" GL (Preto)</Portuguese>
+            <Japanese>HK416A5 10" GL (?)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_01_GL_khk">
+            <English>HK416A5 10" GL (Khaki)</English>
+            <Czech>HK416A5 10" GL (Khaki)</Czech>
+            <French>HK416A5 10" GL (Kaki)</French>
+            <Spanish>HK416A5 10" GL (Caqui)</Spanish>
+            <Russian>HK416A5 10" GL (????)</Russian>
+            <Polish>HK416A5 10" GL (khaki)</Polish>
+            <German>HK416A5 10" GL (Khaki)</German>
+            <Italian>HK416A5 10" GL (Khaki)</Italian>
+            <Hungarian>HK416A5 10" GL (Khaki)</Hungarian>
+            <Portuguese>HK416A5 10" GL (Caqui)</Portuguese>
+            <Japanese>HK416A5 10" GL (??)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_01_GL_snd">
+            <English>HK416A5 10" GL (Sand)</English>
+            <Czech>HK416A5 10" GL (Písková)</Czech>
+            <French>HK416A5 10" GL (Beige)</French>
+            <Spanish>HK416A5 10" GL (Arena)</Spanish>
+            <Russian>HK416A5 10" GL (????????)</Russian>
+            <German>HK416A5 10" GL (Sand)</German>
+            <Polish>HK416A5 10" GL (piaskowy)</Polish>
+            <Italian>HK416A5 10" GL (Sabbia)</Italian>
+            <Hungarian>HK416A5 10" GL (Homok)</Hungarian>
+            <Portuguese>HK416A5 10" GL (Deserto)</Portuguese>
+            <Japanese>HK416A5 10" GL (??)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_02_blk">
+            <English>HK416A5 14.5" (Black)</English>
+            <Czech>HK416A5 14.5" (Cerná)</Czech>
+            <French>HK416A5 14.5" (Noir)</French>
+            <Spanish>HK416A5 14.5" (Negro)</Spanish>
+            <Russian>HK416A5 14.5" (??????)</Russian>
+            <Polish>HK416A5 14.5" (czarny)</Polish>
+            <German>HK416A5 14.5" (Schwarz)</German>
+            <Italian>HK416A5 14.5" (Nero)</Italian>
+            <Hungarian>HK416A5 14.5" (Fekete)</Hungarian>
+            <Portuguese>HK416A5 14.5" (Preto)</Portuguese>
+            <Japanese>HK416A5 14.5" (?)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_02_khk">
+            <English>HK416A5 14.5" (Khaki)</English>
+            <Czech>HK416A5 14.5" (Khaki)</Czech>
+            <French>HK416A5 14.5" (Kaki)</French>
+            <Spanish>HK416A5 14.5" (Caqui)</Spanish>
+            <Russian>HK416A5 14.5" (????)</Russian>
+            <Polish>HK416A5 14.5" (khaki)</Polish>
+            <German>HK416A5 14.5" (Khaki)</German>
+            <Italian>HK416A5 14.5" (Khaki)</Italian>
+            <Hungarian>HK416A5 14.5" (Khaki)</Hungarian>
+            <Portuguese>HK416A5 14.5" (Caqui)</Portuguese>
+            <Japanese>HK416A5 14.5" (??)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_02_snd">
+            <English>HK416A5 14.5" (Sand)</English>
+            <Czech>HK416A5 14.5" (Písková)</Czech>
+            <French>HK416A5 14.5" (Beige)</French>
+            <Spanish>HK416A5 14.5" (Arena)</Spanish>
+            <Russian>HK416A5 14.5" (????????)</Russian>
+            <German>HK416A5 14.5" (Sand)</German>
+            <Polish>HK416A5 14.5" (piaskowy)</Polish>
+            <Italian>HK416A5 14.5" (Sabbia)</Italian>
+            <Hungarian>HK416A5 14.5" (Homok)</Hungarian>
+            <Portuguese>HK416A5 14.5" (Deserto)</Portuguese>
+            <Japanese>HK416A5 14.5" (??)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_03_blk">
+            <English>HK417A2 20" (Black)</English>
+            <Czech>HK417A2 20" (Cerná)</Czech>
+            <French>HK417A2 20" (Noir)</French>
+            <Spanish>HK417A2 20" (Negro)</Spanish>
+            <Russian>HK417A2 20" (??????)</Russian>
+            <Polish>HK417A2 20" (czarny)</Polish>
+            <German>HK417A2 20" (Schwarz)</German>
+            <Italian>HK417A2 20" (Nero)</Italian>
+            <Hungarian>HK417A2 20" (Fekete)</Hungarian>
+            <Portuguese>HK417A2 20" (Preto)</Portuguese>
+            <Japanese>HK417A2 20" (?)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_03_khk">
+            <English>HK417A2 20" (Khaki)</English>
+            <Czech>HK417A2 20" (Khaki)</Czech>
+            <French>HK417A2 20" (Kaki)</French>
+            <Spanish>HK417A2 20" (Caqui)</Spanish>
+            <Russian>HK417A2 20" (????)</Russian>
+            <Polish>HK417A2 20" (khaki)</Polish>
+            <German>HK417A2 20" (Khaki)</German>
+            <Italian>HK417A2 20" (Khaki)</Italian>
+            <Hungarian>HK417A2 20" (Khaki)</Hungarian>
+            <Portuguese>HK417A2 20" (Caqui)</Portuguese>
+            <Japanese>HK417A2 20" (??)</Japanese>
+        </Key>
+        <Key ID="STR_ARK_arifle_SPAR_03_snd">
+            <English>HK417A2 20" (Sand)</English>
+            <Czech>HK417A2 20" (Písková)</Czech>
+            <French>HK417A2 20" (Beige)</French>
+            <Spanish>HK417A2 20" (Arena)</Spanish>
+            <Russian>HK417A2 20" (????????)</Russian>
+            <German>HK417A2 20" (Sand)</German>
+            <Polish>HK417A2 20" (piaskowy)</Polish>
+            <Italian>HK417A2 20" (Sabbia)</Italian>
+            <Hungarian>HK417A2 20" (Homok)</Hungarian>
+            <Portuguese>HK417A2 20" (Deserto)</Portuguese>
+            <Japanese>HK417A2 20'' (??)</Japanese>
+        </Key>
+    </Package>
+</Project>
